+++ internal/config/config_test.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/gkampitakis/go-snaps/snaps"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 	"go.yaml.in/yaml/v3"
 
@@ -166,7 +165,7 @@ func TestSetDisabledChecks(t *testing.T) {
 }
 
 func newRule(t *testing.T, content string) parser.Rule {
-	p := parser.NewParser(false, parser.PrometheusSchema, model.UTF8Validation)
+	p := parser.NewParser(parser.DefaultOptions)
 	file := p.Parse(strings.NewReader(content))
 	if file.Error.Err != nil {
 		t.Error(file.Error)