+++ internal/parser/models_test.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 
 	"github.com/cloudflare/pint/internal/diags"
@@ -13,7 +12,7 @@ import (
 )
 
 func newMustRule(content string) parser.Rule {
-	p := parser.NewParser(false, parser.PrometheusSchema, model.UTF8Validation)
+	p := parser.NewParser(parser.DefaultOptions)
 	file := p.Parse(strings.NewReader(content))
 	if file.Error.Err != nil {
 		panic(file.Error.Err)