+++ internal/reporter/checkstyle_test.go
@@ -8,7 +8,6 @@ import (
 	"testing"
 
 	"github.com/neilotoole/slogt"
-	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 
 	"github.com/cloudflare/pint/internal/checks"
@@ -32,7 +31,7 @@ func TestCheckstyleReporter(t *testing.T) {
 		summary     reporter.Summary
 	}
 
-	p := parser.NewParser(false, parser.PrometheusSchema, model.UTF8Validation)
+	p := parser.NewParser(parser.DefaultOptions)
 	mockFile := p.Parse(strings.NewReader(`
 - record: target is down
   expr: up == 0