+++ internal/reporter/gitlab_test.go
@@ -16,7 +16,6 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/neilotoole/slogt"
-	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 	gitlab "gitlab.com/gitlab-org/api/client-go"
 	"go.nhat.io/httpmock"
@@ -46,7 +45,7 @@ func TestGitLabReporterBadBaseURI(t *testing.T) {
 func TestGitLabReporter(t *testing.T) {
 	type errorCheck func(err error) error
 
-	p := parser.NewParser(false, parser.PrometheusSchema, model.UTF8Validation)
+	p := parser.NewParser(parser.DefaultOptions)
 	mockRules := `
 - record: target is down
   expr: up == 0
@@ -1213,7 +1212,7 @@ func TestGitLabReporterCommentLine(t *testing.T) {
 		showDuplicates  bool
 	}
 
-	p := parser.NewParser(false, parser.PrometheusSchema, model.UTF8Validation)
+	p := parser.NewParser(parser.DefaultOptions)
 	mockFile := p.Parse(strings.NewReader(`
 - record: target is down
   expr: up == 0