+++ docs/changelog.md
@@ -2,6 +2,10 @@
 
 ## v0.80.0
 
+### Added
+
+- Added [promql/features](checks/promql/features.md) check.
+
 ### Fixed
 
 - [promql/vector_matching](checks/promql/vector_matching.md) check no longer reports problems that are also reported by [promql/impossible](checks/promql/impossible.md).